+++ app/(pages)/_styles/day1.module.scss
@@ -1,4 +1,4 @@
-@use '/app/(pages)/mixins.scss';
+@use '/app/(pages)/_styles/mixins.scss';
 
 .main {
     --top-height: 400px;
