+++ app/(pages)/_styles/day2.module.scss
@@ -1,4 +1,4 @@
-@use '/app/(pages)/mixins.scss';
+@use '/app/\(pages\)/_styles/mixins.scss';
 
 .main {
     padding: 40px;
