+++ app/(pages)/_styles/globals.scss
