+++ app/(pages)/_styles/work_card.module.scss
@@ -1,4 +1,4 @@
-@use '/app/(pages)/mixins.scss';
+@use '/app/(pages)/_styles/mixins.scss';
 
 .card {
     --frost-background: rgba(255, 255, 255, 0.1);
