+++ app/(pages)/layout.jsx
@@ -1,5 +1,5 @@
 import { Inter } from "next/font/google";
-import "./globals.scss";
+import "./_styles/globals.scss";
 
 const inter = Inter({ subsets: ["latin"] });
